Moving deploy and pod controllers to another bin
In the future I would like to be able to run only the tag controller and
not the deployment and pod controllers.

This is a step in this direction.
