--- .github/workflows/build.yml
@@ -1,26 +1,0 @@
----
-name: build
-
-on:
-  push:
-    branches:
-      - main
-  pull_request:
-    branches:
-      - main
-
-jobs:
-  build:
-    name: build
-    runs-on: ubuntu-latest
-    container:
-      image: quay.io/tagger/actions-image:latest
-    steps:
-    - name: check out code
-      uses: actions/checkout@v2
-
-    - name: build tagger
-      run: make tagger
-
-    - name: build plugin
-      run: make kubectl-tag
