--- .github/workflows/image.yml
@@ -1,30 +1,0 @@
----
-name: image
-
-on:
-  push:
-    branches:
-      - main
-
-jobs:
-  docker:
-    name: image build and push
-    runs-on: ubuntu-latest
-    steps:
-      - name: check out code
-        uses: actions/checkout@v2
-
-      - name: registry login
-        uses: docker/login-action@v1 
-        with:
-          registry: quay.io
-          username: ${{ secrets.REGISTRY_USERNAME }}
-          password: ${{ secrets.REGISTRY_PASSWORD }}
-
-      - name: build image and push
-        uses: docker/build-push-action@v2
-        with:
-          push: true
-          context: ./
-          file: ./Containerfile
-          tags: quay.io/tagger/operator:latest
