--- .github/workflows/lint.yml
@@ -1,23 +1,0 @@
----
-name: lint
-
-on:
-  push:
-    branches:
-      - main
-  pull_request:
-    branches:
-      - main
-
-jobs:
-  lint:
-    name: lint
-    runs-on: ubuntu-latest
-    container:
-      image: quay.io/tagger/actions-image:latest
-    steps:
-    - name: check out code
-      uses: actions/checkout@v2
-
-    - name: run linter
-      run: golint -set_exit_status ./cmd/kubectl-tag ./cmd/tagger ./controllers/... ./services/...
