+++ .github/workflows/main.yaml
@@ -1,5 +1,5 @@
 ---
-name: release
+name: main
 
 on:
   push:
@@ -7,6 +7,28 @@ on:
       - main
 
 jobs:
+  image:
+    name: image build and push
+    runs-on: ubuntu-latest
+    steps:
+      - name: check out code
+        uses: actions/checkout@v2
+
+      - name: registry login
+        uses: docker/login-action@v1 
+        with:
+          registry: quay.io
+          username: ${{ secrets.REGISTRY_USERNAME }}
+          password: ${{ secrets.REGISTRY_PASSWORD }}
+
+      - name: build image and push
+        uses: docker/build-push-action@v2
+        with:
+          push: true
+          context: ./
+          file: ./Containerfile
+          tags: quay.io/tagger/operator:latest
+
   release:
     name: tagger development release
     runs-on: ubuntu-latest
@@ -15,8 +37,6 @@ jobs:
     steps:
       - name: checkout source code
         uses: actions/checkout@v2
-        with:
-          fetch-depth: 0
 
       - name: build linux plugin
         run: make kubectl-tag
