--- .github/workflows/stable-image.yml
@@ -1,37 +1,0 @@
----
-name: stable image
-
-on:
-  push:
-    tags:
-      - "v*"
-
-jobs:
-  docker:
-    name: image build and push
-    runs-on: ubuntu-latest
-    steps:
-
-      - name: check out code
-        uses: actions/checkout@v2
-
-      - name: registry login
-        uses: docker/login-action@v1 
-        with:
-          registry: quay.io
-          username: ${{ secrets.REGISTRY_USERNAME }}
-          password: ${{ secrets.REGISTRY_PASSWORD }}
-
-      - name: extract current tag
-        id: get_tag
-        run: echo ::set-output name=tag::$(echo $GITHUB_REF | cut -d / -f 3)
-
-      - name: build image and push
-        uses: docker/build-push-action@v2
-        with:
-          push: true
-          build-args:
-            version=${{ steps.get_tag.outputs.tag }}
-          context: ./
-          file: ./Containerfile
-          tags: quay.io/tagger/operator:${{ steps.get_tag.outputs.tag }}
