--- .github/workflows/unit.yml
@@ -1,23 +1,0 @@
----
-name: unit
-
-on:
-  push:
-    branches:
-      - main
-  pull_request:
-    branches:
-      - main
-
-jobs:
-  unit:
-    name: unit
-    runs-on: ubuntu-latest
-    container:
-      image: quay.io/tagger/actions-image:latest
-    steps:
-    - name: check out source code
-      uses: actions/checkout@v2
-
-    - name: run tests
-      run: go test -mod vendor -v ./...
