+++ Containerfile
@@ -14,6 +14,7 @@ ARG version
 ENV VERSION=${version:-v0.0.0}
 COPY . .
 RUN make tagger
+RUN make depctrl
 
 #
 # Application
@@ -21,6 +22,7 @@ RUN make tagger
 FROM docker.io/fedora:latest
 RUN dnf install -y device-mapper-libs
 COPY --from=builder /src/output/bin/tagger /usr/local/bin/tagger
+COPY --from=builder /src/output/bin/depctrl /usr/local/bin/depctrl
 # 8080 pod mutating webhook handler.
 # 8081 quay webhooks handler.
 # 8082 docker webhooks handler.
