+++ chart/templates/deploy-depctrl.yaml
@@ -1,0 +1,32 @@
+---
+apiVersion: apps/v1
+kind: Deployment
+metadata:
+  name: depctrl
+  namespace: {{ .Release.Namespace }}
+  labels:
+    app: depctrl
+spec:
+  replicas: {{ .Values.replicas }}
+  selector:
+    matchLabels:
+      app: depctrl
+  template:
+    metadata:
+      labels:
+        app: depctrl
+    spec:
+      serviceAccountName: tagger
+      containers:
+      - name: depctrl
+        image: {{ .Values.image }}
+        imagePullPolicy: Always
+        env:
+        - name: POD_NAMESPACE
+          valueFrom:
+            fieldRef:
+              fieldPath: metadata.namespace
+        - name: POD_NAME
+          valueFrom:
+            fieldRef:
+              fieldPath: metadata.name
